Fix AutoCop and Run now button
Fix #12
